docs: Updated README.md
Updated change log, and updated the instructions for how to download the postman collection. As doing a right-click, save as within the browser corrupts the collection. By saving it in html format, in which the import will fail. Similar comments of this failure have also been included in the vidcast.